Update build instructions
The instructions now inform users to perform an out-of-tree build, which
keeps build artifacts separate from the source code. This makes it
easier to ensure non-source code is properly gitignored, and to more
easily clean it up afterwards.

I also added an instruction on how to prepare the software for packaging
(ie. in a Linux distro).